change

See merge request heo3793/pjt-test!15

+++ final-pjt-back/communitys/models.py
@@ -9,7 +9,7 @@ class Review(models.Model):
     title = models.CharField(max_length=30)
     content = models.TextField()
     movie_title = models.CharField(max_length=50)
-    score = models.IntegerField(default=0, validators=[MinValueValidator(0),MaxValueValidator(5)])
+    score = models.FloatField(default=0, validators=[MinValueValidator(0),MaxValueValidator(5)])
     created_at = models.DateTimeField(auto_now_add = True)
     updated_at = models.DateTimeField(auto_now= True)
     like_users = models.ManyToManyField(settings.AUTH_USER_MODEL, related_name='like_reviews')
